Secure_Password_Manager/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
pycryptodome==3.19.0
zxcvbn==4.4.28
requests==2.31.0
python-dotenv==1.0.0
